+++ app/src/main/java/com/example/foody/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.foody
+package com.example.foody.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -7,8 +7,8 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
+import com.example.foody.R
 import com.example.foody.databinding.ActivityMainBinding
-import com.example.foody.databinding.FragmentRecipesBinding
 
 class MainActivity : AppCompatActivity() {
     //view Binding of activity_main layout
@@ -26,9 +26,9 @@ class MainActivity : AppCompatActivity() {
         //setting up bottomNavigationView
         val navController = findNavController(R.id.navHostFragment)
         val appBarConfiguration = AppBarConfiguration(setOf(
-                R.id.recipesFragment,
-                R.id.favoriteRecipesFragment,
-                R.id.foodJokeFragment
+            R.id.recipesFragment,
+            R.id.favoriteRecipesFragment,
+            R.id.foodJokeFragment
         ))
 
         //Accessing views from activity_main layout
