+++ app/src/main/res/layout/fragment_favorite_recipes.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/lightGray"
-    tools:context=".FavoriteRecipesFragment">
+    tools:context=".ui.fragments.favorites.FavoriteRecipesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
