+++ app/src/main/res/layout/fragment_food_joke.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/lightGray"
-    tools:context=".FoodJokeFragment">
+    tools:context=".ui.fragments.foodjoke.FoodJokeFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
